rename: 파일 이름들에 네이밍 컨벤션 적용
